notification icon improvements
