--- src/components/Notification/Notification.css
@@ -1,11 +1,0 @@
-.notification-icon {
-    background-color: red;
-    height: 23px;
-    width: 23px;
-    color: white;
-    border-radius: 50%;
-    float: right;
-    display: block;
-    margin: 0 20px;
-    text-align: center
-}
