--- src/components/Notification/Notification.tsx
@@ -1,17 +1,0 @@
-import './Notification.css';
-
-export interface NotificationIconProps {
-    count: number;
-}
-
-const NotificationIcon = (props: NotificationIconProps) => {
-
-    return (
-        <span className='notification-icon'>
-            {props.count}
-        </span>
-    )
-
-}
-
-export default NotificationIcon;
