--- src/components/Notification/index.ts
@@ -1,1 +1,0 @@
-export * from './Notification';
