+++ src/components/NotificationIcon/NotificationIcon.css
@@ -1,0 +1,21 @@
+.notification {
+    border-radius: 50%;
+    display: block;
+    position:relative;
+    height: 20px;
+    width: 20px;
+  }
+  
+  .notification-count {
+    align-items: center;
+    background-size: cover;
+    border-radius: 50%;
+    color: #fff;
+    display: flex;
+    font-size: var(--font-tiny);
+    font-weight: 500;
+    height: 100%;
+    justify-content: center;
+    line-height: 120%;
+    width: 100%;
+  }
