+++ src/components/NotificationIcon/index.ts
@@ -1,0 +1,1 @@
+export * from './NotificationIcon';
