+++ src/i18n/en/dashboard-sidebar.json
@@ -4,6 +4,8 @@
   "All projects": "All projects",
   "Account": "Account",
   "Preferences": "Preferences",
+  "Account Preferences": "Account Preferences",
+  "Notifications": "Notifications",
   "Documentation": "Documentation",
   "Tutorial": "Tutorial",
   "FAQ": "FAQ",
